+++ app/config/config.go
@@ -43,6 +43,12 @@ orderKeeperConcurrency = {{ .BaseConfig.OrderKeeperConcurrency }}
 # Days count back for breathe block
 breatheBlockDaysCountBack = {{ .BaseConfig.BreatheBlockDaysCountBack }}
 
+[upgrade]
+# fixOrderSeqInPriceLevel height
+fixOrderSeqInPriceLevelHeight = {{ .UpgradeConfig.FixOrderSeqInPriceLevelHeight }}
+# fixDropFilledOrderSeq height
+fixDropFilledOrderSeqHeight = {{ .UpgradeConfig.FixDropFilledOrderSeqHeight }}
+
 [addr]
 # Bech32PrefixAccAddr defines the Bech32 prefix of an account's address
 bech32PrefixAccAddr = "{{ .AddressConfig.Bech32PrefixAccAddr }}"
@@ -130,6 +136,7 @@ type BinanceChainConfig struct {
 	*PublicationConfig `mapstructure:"publication"`
 	*LogConfig         `mapstructure:"log"`
 	*BaseConfig        `mapstructure:"base"`
+	*UpgradeConfig     `mapstructure:"upgrade"`
 }
 
 func DefaultBinanceChainConfig() *BinanceChainConfig {
@@ -138,6 +145,7 @@ func DefaultBinanceChainConfig() *BinanceChainConfig {
 		PublicationConfig: defaultPublicationConfig(),
 		LogConfig:         defaultLogConfig(),
 		BaseConfig:        defaultBaseConfig(),
+		UpgradeConfig:     defaultUpgradeConfig(),
 	}
 }
 
@@ -272,6 +280,18 @@ func defaultBaseConfig() *BaseConfig {
 		BreatheBlockInterval:      0,
 		OrderKeeperConcurrency:    2,
 		BreatheBlockDaysCountBack: 7,
+	}
+}
+
+type UpgradeConfig struct {
+	FixOrderSeqInPriceLevelHeight int64
+	FixDropFilledOrderSeqHeight   int64
+}
+
+func defaultUpgradeConfig() *UpgradeConfig {
+	return &UpgradeConfig{
+		FixDropFilledOrderSeqHeight:   2855000,
+		FixOrderSeqInPriceLevelHeight: 2855000,
 	}
 }