+++ common/utils/numbers.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"encoding/binary"
-	"math/big"
 )
 
 func AbsInt(a int64) int64 {
@@ -24,22 +23,21 @@ func MaxInt(a, b int64) int64 {
 	return b
 }
 
-// CalBigNotional() calculate the multiply value of notional based on price and qty
-// both price and qty are in int64 with 1e8 as decimals
-// TODO: here the floor divide is used. there may cause small residual.
-func CalBigNotional(price, qty int64) int64 {
-	var bi big.Int
-	return bi.Div(bi.Mul(big.NewInt(qty), big.NewInt(price)), big.NewInt(1e8)).Int64()
-}
-
-// IsExceedMaxNotional return the result that is the product of price and quantity exceeded max notional
-func IsExceedMaxNotional(price, qty int64) bool {
-	var bi big.Int
-	return !bi.Div(bi.Mul(big.NewInt(qty), big.NewInt(price)), big.NewInt(1e8)).IsInt64()
-}
-
 func Int642Bytes(n int64) []byte {
 	b := make([]byte, 8)
 	binary.LittleEndian.PutUint64(b, uint64(n))
 	return b
+}
+
+func Mul64(a, b int64) (int64, bool) {
+	if a == 0 || b == 0 {
+		return 0, true
+	}
+	c := a * b
+	if (c < 0) == ((a < 0) != (b < 0)) {
+		if c/b == a {
+			return c, true
+		}
+	}
+	return c, false
 }