+++ plugins/dex/list/hooks_test.go
@@ -5,11 +5,12 @@ import (
 	"testing"
 	"time"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/gov"
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/log"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/gov"
 
 	"github.com/binance-chain/node/common/types"
 	dexTypes "github.com/binance-chain/node/plugins/dex/types"
@@ -81,6 +82,28 @@ func TestQuoteAssetEmpty(t *testing.T) {
 	require.NotNil(t, err, "err should not be nil")
 
 	require.Contains(t, err.Error(), "quote asset symbol should not be empty")
+}
+
+func TestEqualBaseAssetAndQuoteAsset(t *testing.T) {
+	hooks := NewListHooks(nil, nil)
+
+	listParams := gov.ListTradingPairParams{
+		BaseAssetSymbol:  "BNB",
+		QuoteAssetSymbol: "BNB",
+	}
+
+	listParamsBz, err := json.Marshal(listParams)
+	require.Nil(t, err, "marshal list params error")
+
+	proposal := gov.TextProposal{
+		ProposalType: gov.ProposalTypeListTradingPair,
+		Description:  string(listParamsBz),
+	}
+
+	err = hooks.OnProposalSubmitted(sdk.Context{}, &proposal)
+	require.NotNil(t, err, "err should not be nil")
+
+	require.Contains(t, err.Error(), "base token and quote token should not be the same")
 }
 
 func TestWrongPrice(t *testing.T) {
@@ -221,7 +244,7 @@ func TestPrerequisiteTradingPair(t *testing.T) {
 
 	err = hooks.OnProposalSubmitted(ctx, &proposal)
 	require.NotNil(t, err, "err should not be nil")
-	require.Contains(t, err.Error(), "Token BTC-ABC should be listed against BNB before against ETH-ABC")
+	require.Contains(t, err.Error(), "token BTC-ABC should be listed against BNB before against ETH-ABC")
 
 	pair := dexTypes.NewTradingPair(listParams.BaseAssetSymbol, types.NativeTokenSymbol, listParams.InitPrice)
 	err = orderKeeper.PairMapper.AddTradingPair(ctx, pair)
@@ -229,7 +252,7 @@ func TestPrerequisiteTradingPair(t *testing.T) {
 
 	err = hooks.OnProposalSubmitted(ctx, &proposal)
 	require.NotNil(t, err, "err should not be nil")
-	require.Contains(t, err.Error(), "Token ETH-ABC should be listed against BNB before listing BTC-ABC against ETH-ABC")
+	require.Contains(t, err.Error(), "token ETH-ABC should be listed against BNB before listing BTC-ABC against ETH-ABC")
 
 	pair = dexTypes.NewTradingPair(listParams.QuoteAssetSymbol, types.NativeTokenSymbol, listParams.InitPrice)
 	err = orderKeeper.PairMapper.AddTradingPair(ctx, pair)