+++ plugins/dex/store/mapper_test.go
@@ -56,7 +56,7 @@ func TestMapper_GetAddTradingPair(t *testing.T) {
 	require.NotEmpty(t, pair)
 	require.Equal(t, baseAsset, pair.BaseAssetSymbol)
 	require.Equal(t, quoteAsset, pair.QuoteAssetSymbol)
-	require.Equal(t, utils.Fixed8(1e8), pair.Price)
+	require.Equal(t, utils.Fixed8(1e8), pair.ListPrice)
 	require.Equal(t, utils.Fixed8(1), pair.TickSize)
 	require.Equal(t, utils.Fixed8(1e8), pair.LotSize)
 }